requirements.txt
numpy==1.26.4       # Critical: Avoid 2.0+ to keep compatibility with old math
scipy==1.12.0       # Stable for signal filters
sounddevice==0.4.6  # Reliable C-interfacing for audio
pythonnet==3.0.3    # For the UIAL C# Bridge
phonemizer==3.2.1   # For Kokoro frontend
sounddevice
soundfile
pyworld-prebuilt
librosa
gruut
pandas
spacy 
gruut[fr]
matplotlib
soundfile
pykakasi